+++ examples/HirePurchaseDirectDebit/Constants.php
@@ -1,6 +1,6 @@
 <?php
 /**
- * This file defines the constants needed for the Invoice Factoring example.
+ * This file defines the constants needed for the hire purchase direct debit example.
  *
  * Copyright (C) 2019 heidelpay GmbH
  *
@@ -25,5 +25,6 @@
 
 require_once __DIR__ . '/../Constants.php';
 
-define('EXAMPLE_URL', EXAMPLE_BASE_FOLDER . 'InvoiceFactoring');
+define('EXAMPLE_URL', EXAMPLE_BASE_FOLDER . 'HirePurchaseDirectDebit');
 define('CONTROLLER_URL', EXAMPLE_URL . '/Controller.php');
+define('CONFIRM_URL', EXAMPLE_URL . '/confirm.php');