+++ CHANGELOG.md
@@ -6,6 +6,7 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/) a
 ## [1.2.5.0][1.2.5.0]
 
 ### Added
+*   Hire purchase payment type (FlexiPay Rate).
 *   Property invoiceId to `Payment` class.
 
 ### Changed